loadOlder and loadNewer improvements in useStream hook
They only update from the previous event and next event respectively.
This avoids issues where if the previous event occurred too
far in the past or the next event occurs too far in the future from the
current stream boundary, the API returns a 500 error saying that the
requested window is too long.